requirements.txt
blinker==1.9.0
Bottleneck==1.4.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mysql-connector-python==9.0.0
numexpr==2.10.1
numpy==1.26.4
pandas==2.2.3
pip==25.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2023.3
Werkzeug==3.1.3
wheel==0.45.1
